Allow user-specified provisioners, make class members private.
User-specific provisioners are now exposed as a template parameter in the
parse() functions. This allows users to directly manage memory allocations,
e.g., to create R/Python objects without an intermediate C++ object.

The object and array provisioner methods now accept a fully formed vector/map
of elements. This allows us to remove various methods (e.g., add()) and
simplifies the provisioner contract as the user is not obliged to implement any
methods on the classes (as each instance is not used after construction).

While we're making breaking changes, we also:

- Remove the quick getters, which are more dangerous than convenient.
- Make class members private, to align with coding style in the tatami guide.
- Remove the old file opening functions that don't accept Options.