Optionally preserve the string representation for numbers.
This is inspired by a similar feature in Golang. It is occasionally
useful for faithful roundtrips to avoid loss of precision, or if the
string representation itself is needed for, e.g., text searches.

We introduce a ParseOptions argument to control this behavior, and we
add overloads to the validate/parse functions to pass options. We
soft-deprecate the original functions for simplicity.